Removed comments with TODOs because they were already fixed.
Overwriting the prompt and going into the negative place were fixed by
removing the removed character from the buffer. The handleInput function
will only run if the buffer has something in it so it will not erase the
prompt or keep moving past it.

There are still some glitches (like with underscore) but I believe these
are caused by issues with the font and the font engine and not how
backspacing is handled.