Fixed code style. Fixed comments.
Added a variable for start of line instead of hard coded 0. I removed
comments that were clutter and added some smaller and more useful ones.
I changed tabs to 2 because that is what I prefer and that is all that
matters.

Look at the compiled JS to get a feel for the few code changes I made
since the indenting is not changed in there.