Added command history.
This a a complete fuckery. Edited keyboard driver to return ANSI control
codes for arrow keys. Edited console to call handler functions for those
command keys. Added handler functions to do the command history with
those command keys.

A massive refactoring is required for console and shell because this is
awful but I got really god damn annoyed with this fucking shit so I
couldn't be bothered at this point in time.

See #5.