Fixes #69, problems with tab completion.
Commit 17f5ecb broke tab completion such that the last letter you
entered got deleted by the tab completion. This was due to a change such
that the shell was capturing tabs before they got sent to the display.
Before the tab got displayed and the shell had to delete it with a
backspace first. Since the tab stopped getting displayed the backspace
was not needed and is now removed.