Ok. Fuck references.
So my memory is made up of references. A program set the same byte in
memory to multiple other places in memory. So it set the SAME reference
to the SAME byte to DIFFERENT spots in memory. So when you changed one,
it changed them all. So now ALL memory in the CPU creates NEW bytes.
Fuck that shit bro.