Quick fix for BNE.
Alan told me that his test program was broken. Traced the error to
System call 1. It uses Stdio.putStr, which assumes it is called by the
shell so it does not return to the user program correctly. It leaves the
CPU in kernel mode so the BNE instruction does not wrap the address
around. This fix simply disables kernel mode for BNE so it will always
wrap the address.